Modernize library
Library has not been maintained in a long time, and no longer compiles
on newest Rust versions. Update the library to work in newest Rust (1.14
as of this writing.)

The changes here do not change the workings of the library, with the
major exception of changing peek_byte's return type from `IoResult<u8>`
(which equivalently is) `Result<u8, io::Error>`, into
`Result<u8, &io::Error>`, as io::Error is no longer copyable.

None of the tests are changed, they all still need updating.